requirements.txt
python-telegram-bot==20.3
yt-dlp==2026.3.3
python-dotenv==1.0.0
